move UpdatingInstructions to simulation_instruction.py and rename to instructions.py
