+++ examples/control_new_events.py
@@ -3,7 +3,7 @@ from lukefi.metsi.domain.forestry_types import ForestCondition
 from lukefi.metsi.domain.natural_processes.grow_motti_dll import grow_motti_dll_fn
 from lukefi.metsi.domain.pre_ops import generate_reference_trees
 from lukefi.metsi.sim.transition import Transition
-from lukefi.metsi.sim.simulation_instruction import SimulationInstruction
+from lukefi.metsi.sim.instructions import SimulationInstruction
 from examples.declarations.sqlite import sqlite_decl
 
 from user_events import Mounding
