+++ examples/control_updating_treatments.py
@@ -1,14 +1,6 @@
-from lukefi.metsi.domain.collected_data import NaturalProcessInfo
-from lukefi.metsi.domain.conditions import RelativeTimePoints
-from lukefi.metsi.domain.forestry_types import ForestCondition
-from lukefi.metsi.domain.natural_processes.updating import update_to_year
 from lukefi.metsi.domain.natural_processes.grow_acta import grow_acta_fn
 from lukefi.metsi.domain.pre_ops import filter_trees, generate_reference_trees, filter_stands, scale_area_weight
-from lukefi.metsi.sim.generators import Alternatives, Event, Sequence
-from lukefi.metsi.sim.sim_configuration import UpdatingInstructions
-from lukefi.metsi.sim.transition import Transition
-from lukefi.metsi.sim.simulation_instruction import SimulationInstruction
-from lukefi.metsi.sim.treatment import do_nothing
+from lukefi.metsi.sim.instructions import UpdatingInstructions
 from examples.declarations.export_prepro import mela_decl
 from examples.declarations.sqlite import sqlite_decl
