+++ examples/optional_generator.py
@@ -5,7 +5,7 @@ from lukefi.metsi.domain.natural_processes.grow_acta import grow_acta_fn
 from lukefi.metsi.domain.pre_ops import filter_trees, generate_reference_trees, filter_stands, scale_area_weight
 from lukefi.metsi.sim.generators import Alternatives, Event, Optional, Sequence
 from lukefi.metsi.sim.transition import Transition
-from lukefi.metsi.sim.simulation_instruction import SimulationInstruction
+from lukefi.metsi.sim.instructions import SimulationInstruction
 from lukefi.metsi.sim.treatment import do_nothing
 from examples.declarations.sqlite import sqlite_decl
