+++ tests/resources/runners_test/parameters_branching.py
@@ -1,6 +1,6 @@
 from lukefi.metsi.domain.conditions import TimePoints
 from lukefi.metsi.sim.condition import Condition
-from lukefi.metsi.sim.simulation_instruction import SimulationInstruction
+from lukefi.metsi.sim.instructions import SimulationInstruction
 from lukefi.metsi.sim.generators import Alternatives, Sequence, Event
 from lukefi.metsi.sim.treatment import do_nothing
 from tests.toy_model import ToyModel, ToyTransition, toy_inc
