+++ tests/robot/MottiGrowTransition/input/control.py
@@ -11,7 +11,7 @@ from lukefi.metsi.domain.pre_ops import (
 from lukefi.metsi.domain.events import DoNothing
 from lukefi.metsi.sim.condition import Condition
 from lukefi.metsi.sim.transition import Transition
-from lukefi.metsi.sim.simulation_instruction import SimulationInstruction
+from lukefi.metsi.sim.instructions import SimulationInstruction
 
 
 control_structure = {
