+++ tests/sim/sim_configuration_test.py
@@ -4,7 +4,7 @@ from lukefi.metsi.domain.conditions import TimeSinceTreatment
 from lukefi.metsi.sim.condition import Condition
 from lukefi.metsi.sim.generators import Event, Sequence
 from lukefi.metsi.sim.sim_configuration import SimConfiguration
-from lukefi.metsi.sim.simulation_instruction import SimulationInstruction
+from lukefi.metsi.sim.instructions import SimulationInstruction
 from lukefi.metsi.sim.treatment import Treatment
 from tests.toy_model import ToyModel, ToyTransition
