+++ tests/sim/simulation_instructions_test.py
@@ -1,11 +1,10 @@
 import sqlite3
 import unittest
 
-from lukefi.metsi.domain.utils.file_io import create_database_tables
 from lukefi.metsi.sim.condition import Condition
 from lukefi.metsi.sim.generators import Alternatives, Event
 from lukefi.metsi.sim.sim_configuration import SimConfiguration
-from lukefi.metsi.sim.simulation_instruction import SimulationInstruction
+from lukefi.metsi.sim.instructions import SimulationInstruction
 from lukefi.metsi.sim.simulation_payload import SimulationPayload
 from lukefi.metsi.sim.simulator import _simulate_unit
 from tests.toy_model import ToyModel, ToyTransition, toy_inc
